move cache implementations from engine to spark project
