--- spark/src/main/scala/geotrellis/spark/io/Cache.scala
@@ -1,7 +1,0 @@
-package geotrellis.spark.io
-
-trait Cache[K, V] {
-  def contains(key: K): Boolean
-  def apply(key: K): Option[V]
-  def update(key: K, value: V): Unit
-}
